backend/requirements.txt
fastapi[all]
pyserial